+++ react/components/ProductPrice/index.js
@@ -24,26 +24,7 @@ class Price extends Component {
     showLabels: true,
     showInstallments: false,
     showSavings: false,
-    labelSellingPrice: null,
-    installmentsClasses: {},
-    classes: {
-      root: null,
-      rootLoader: null,
-      listPrice: {
-        container: null,
-        label: null,
-        value: null
-      },
-      sellingPrice: {
-        container: null,
-        label: null,
-        value: null
-      },
-      savings: {
-        container: null,
-        value: null
-      }
-    }
+    labelSellingPrice: null
   }
 
   currencyOptions = {
@@ -62,46 +43,51 @@ class Price extends Component {
       showLabels,
       showSavings,
       labelSellingPrice,
+      className,
+      loaderClass,
+      listPriceContainerClass,
+      listPriceLabelClass,
+      listPriceClass,
+      sellingPriceContainerClass,
+      sellingPriceLabelClass,
+      sellingPriceClass,
+      savingsContainerClass,
+      savingsClass,
       installments,
-      installmentsClasses,
+      installmentClass,
+      interestRateClass,
+      installmentContainerClass,
       styles,
       intl: { formatNumber },
     } = this.props
 
-    let { classes } = this.props
-    // avoiding undefined verifications
-    classes = {
-      ...Price.defaultProps.classes,
-      ...classes
-    }
-
     if ((showListPrice && isNil(listPrice)) || isNil(sellingPrice)) {
-      return <Price.Loader classes={classes} {...styles} />
+      return <Price.Loader loaderClass={loaderClass} {...styles} />
     }
 
     const differentPrices = showListPrice && sellingPrice !== listPrice
 
     return (
-      <div className={classNames('vtex-price', classes.root)}>
+      <div className={classNames('vtex-price', className)}>
         {differentPrices && (
-          <div className={classNames('vtex-price-list__container', classes.listPrice.container)}>
+          <div className={classNames('vtex-price-list__container', listPriceContainerClass)}>
             {showLabels && (
-              <div className={classNames('vtex-price-list__label', classes.listPrice.label)}>
+              <div className={classNames('vtex-price-list__label', listPriceLabelClass)}>
                 <FormattedMessage id="pricing.from" />
               </div>
             )}
-            <span className={classNames('vtex-price-list', classes.listPrice.value)}>
+            <span className={classNames('vtex-price-list', listPriceClass)}>
               {formatNumber(listPrice, this.currencyOptions)}
             </span>
           </div>
         )}
-        <div className={classNames('vtex-price-selling__container', classes.sellingPrice.container)}>
+        <div className={classNames('vtex-price-selling__container', sellingPriceContainerClass)}>
           {showLabels && (
-            <div className={classNames('vtex-price-selling__label', classes.sellingPrice.label)}>
+            <div className={classNames('vtex-price-selling__label', sellingPriceLabelClass)}>
               {labelSellingPrice || <FormattedMessage id="pricing.to" />}
             </div>
           )}
-          <div className={classNames('vtex-price-selling', classes.sellingPrice.value)}>
+          <div className={classNames('vtex-price-selling', sellingPriceClass)}>
             {formatNumber(sellingPrice, this.currencyOptions)}
           </div>
         </div>
@@ -111,11 +97,13 @@ class Price extends Component {
             showLabels={showLabels}
             formatNumber={formatNumber}
             currencyOptions={this.currencyOptions}
-            classes={installmentsClasses}
+            className={installmentContainerClass}
+            interestRateClass={interestRateClass}
+            installmentClass={installmentClass}
           />}
         {differentPrices && showSavings && (
-          <div className={classNames('vtex-price-savings__container', classes.savings.container)}>
-            <div className={classNames('vtex-price-savings', classes.savings.value)}>
+          <div className={classNames('vtex-price-savings__container', savingsContainerClass)}>
+            <div className={classNames('vtex-price-savings', savingsClass)}>
               <FormattedMessage
                 id="pricing.savings"
                 values={{
@@ -134,7 +122,7 @@ class Price extends Component {
 }
 
 Price.Loader = (loaderProps = {}) => (
-  <div className={classNames('vtex-price vtex-price-loader', loaderProps.classes.rootLoader)}>
+  <div className={classNames('vtex-price vtex-price-loader', loaderProps.loaderClass)}>
     <ContentLoader
       style={{
         width: '100%',